Separated router 'files' (images) from vehicle's files
