+++ main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from routers import bookings_router, payments_router, reviews_router, users_router, vehicles_router
+from routers import bookings_router, payments_router, reviews_router, users_router, vehicles_router, files_router
 from auth import authentication
 from db import models
 from db.database import engine
@@ -13,6 +13,7 @@ app.include_router(reviews_router.router)
 app.include_router(users_router.router)
 app.include_router(vehicles_router.router)
 app.include_router(authentication.router)
+app.include_router(files_router.router)
 
 models.Base.metadata.create_all(engine)
